+++ packages/apps/viewer-sample-react/src/providers/TestUiItemsProvider2.tsx
@@ -1,0 +1,38 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
+
+import {
+  AbstractWidgetProps,
+  StagePanelLocation,
+  StagePanelSection,
+  UiItemsProvider,
+} from "@bentley/ui-abstract";
+import { FillCentered } from "@bentley/ui-core";
+import React from "react";
+
+export class TestUiProvider2 implements UiItemsProvider {
+  public readonly id = "TestUiProvider2";
+
+  public provideWidgets(
+    stageId: string,
+    _stageUsage: string,
+    location: StagePanelLocation,
+    _section?: StagePanelSection | undefined
+  ): ReadonlyArray<AbstractWidgetProps> {
+    const widgets: AbstractWidgetProps[] = [];
+    if (
+      stageId === "DefaultFrontstage" &&
+      location === StagePanelLocation.Right
+    ) {
+      widgets.push({
+        id: "addonWidget",
+        getWidgetContent: () => (
+          <FillCentered>Addon Widget in panel</FillCentered>
+        ),
+      });
+    }
+    return widgets;
+  }
+}