+++ packages/apps/viewer-sample-react/src/providers/index.ts
@@ -3,8 +3,5 @@
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
 
-export * from "./AuthClientHome";
-export * from "./Home";
-export * from "./UserManagerHome";
-export * from "./Header";
-export * from "./BlankConnectionHome";
+export * from "./TestUiItemsProvider";
+export * from "./TestUiItemsProvider2";