+++ packages/modules/itwin-viewer-react/src/hooks/index.ts
@@ -3,8 +3,6 @@
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
 
-export * from "./AuthClientHome";
-export * from "./Home";
-export * from "./UserManagerHome";
-export * from "./Header";
-export * from "./BlankConnectionHome";
+export * from "./useExtensions";
+export * from "./useTheme";
+export * from "./useUiProviders";