+++ packages/modules/itwin-viewer-react/src/hooks/useTheme.tsx
@@ -3,8 +3,14 @@
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
 
-export * from "./AuthClientHome";
-export * from "./Home";
-export * from "./UserManagerHome";
-export * from "./Header";
-export * from "./BlankConnectionHome";
+import { ColorTheme, UiFramework } from "@bentley/ui-framework";
+import { useEffect } from "react";
+
+export function useTheme(theme?: ColorTheme | string): void {
+  useEffect(() => {
+    if (theme) {
+      // use the provided theme
+      UiFramework.setColorTheme(theme);
+    }
+  }, [theme]);
+}