+++ packages/modules/itwin-viewer-react/src/services/Initializer.ts
@@ -6,7 +6,11 @@
 import { ClientRequestContext } from "@bentley/bentleyjs-core";
 import { Config } from "@bentley/bentleyjs-core";
 import { FrontendApplicationInsightsClient } from "@bentley/frontend-application-insights-client";
-import { BentleyCloudRpcParams } from "@bentley/imodeljs-common";
+import {
+  BentleyCloudRpcParams,
+  RpcInterface,
+  RpcInterfaceDefinition,
+} from "@bentley/imodeljs-common";
 import { IModelApp, IModelAppOptions } from "@bentley/imodeljs-frontend";
 import { I18N } from "@bentley/imodeljs-i18n";
 import { UrlDiscoveryClient } from "@bentley/itwin-client";
@@ -17,13 +21,13 @@ import { UiComponents } from "@bentley/ui-components";
 import { UiCore } from "@bentley/ui-core";
 import {
   AppNotificationManager,
+  ConfigurableUiManager,
   FrameworkReducer,
   FrameworkUiAdmin,
   StateManager,
   UiFramework,
 } from "@bentley/ui-framework";
 
-import { AppUi } from "../components/app-ui/AppUi";
 import { initRpc } from "../config/rpc";
 import { IModelBackendOptions, ItwinViewerInitializerParams } from "../types";
 import { ai, trackEvent } from "./telemetry/TelemetryService";
@@ -32,22 +36,29 @@ import { ai, trackEvent } from "./telemetry/TelemetryService";
 class Initializer {
   private static _initialized: Promise<void>;
   private static _initializing = false;
+  private static _iModelDataErrorMessage: string | undefined;
+  private static _synchronizerRootUrl: string | undefined;
 
   /** initialize rpc */
   private static async _initializeRpc(
     backendOptions?: IModelBackendOptions,
-    isDesktop?: boolean
+    isDesktop?: boolean,
+    additionalRpcInterfaces?: RpcInterfaceDefinition<RpcInterface>[]
   ): Promise<void> {
     // if rpc params for a custom backend are provided, initialized with those
     if (
       backendOptions?.customBackend &&
       backendOptions.customBackend.rpcParams
     ) {
-      return initRpc(backendOptions.customBackend.rpcParams, isDesktop);
+      return initRpc(
+        backendOptions.customBackend.rpcParams,
+        isDesktop,
+        additionalRpcInterfaces
+      );
     }
     const rpcParams = await this._getHostedConnectionInfo(backendOptions);
     if (rpcParams) {
-      return initRpc(rpcParams, isDesktop);
+      return initRpc(rpcParams, isDesktop, additionalRpcInterfaces);
     }
   }
 
@@ -96,9 +107,51 @@ class Initializer {
     }
   }
 
+  public static async getSynchronizerUrl(
+    contextId: string,
+    iModelId: string
+  ): Promise<string> {
+    if (!this._synchronizerRootUrl) {
+      const urlDiscoveryClient = new UrlDiscoveryClient();
+      this._synchronizerRootUrl = await urlDiscoveryClient.discoverUrl(
+        new ClientRequestContext(),
+        "itwinbridgeportal",
+        Config.App.get("imjs_buddi_resolve_url_using_region")
+      );
+    }
+    const portalUrl = `${this._synchronizerRootUrl}/${contextId}/${iModelId}`;
+    return IModelApp.i18n.translateWithNamespace(
+      "iTwinViewer",
+      "iModels.synchronizerLink",
+      {
+        bridgePortal: portalUrl,
+      }
+    );
+  }
+
   /** expose initialized promise */
   public static get initialized(): Promise<void> {
     return this._initialized;
+  }
+
+  /** Message to display when there are iModel data-related errors */
+  public static async getIModelDataErrorMessage(
+    contextId: string,
+    iModelId: string,
+    prefix?: string
+  ): Promise<string> {
+    if (this._iModelDataErrorMessage !== undefined) {
+      return prefix
+        ? `${prefix} ${this._iModelDataErrorMessage}`
+        : this._iModelDataErrorMessage;
+    }
+    const synchronizerPortalUrl = await this.getSynchronizerUrl(
+      contextId,
+      iModelId
+    );
+    return prefix
+      ? `${prefix} ${synchronizerPortalUrl}`
+      : synchronizerPortalUrl;
   }
 
   /** shutdown IModelApp */
@@ -222,12 +275,13 @@ class Initializer {
         // allow uiAdmin to open key-in palette when Ctrl+F2 is pressed - good for manually loading extensions
         IModelApp.uiAdmin.updateFeatureFlags({ allowKeyinPalette: true });
 
-        AppUi.initialize();
+        ConfigurableUiManager.initialize();
 
         // initialize RPC communication
         await Initializer._initializeRpc(
           viewerOptions?.backend,
-          viewerOptions?.desktopApp
+          viewerOptions?.desktopApp,
+          viewerOptions?.additionalRpcInterfaces
         );
 
         if (viewerOptions?.appInsightsKey) {
@@ -237,6 +291,9 @@ class Initializer {
         await PropertyGridManager.initialize(IModelApp.i18n);
 
         await TreeWidget.initialize(IModelApp.i18n);
+
+        // override the defaut daa error message
+        this._iModelDataErrorMessage = viewerOptions?.iModelDataErrorMessage;
 
         console.log("iModel.js initialized");